added sfx for boulder & orbital, updated volume for weapons, added phase bgm
added sfx to boulder and orbital prefabs. changed the audio volume for the various weapons and other audio to 0.1 and less if it was still loud. I also added bgm from the beginning of the game, then changes in run phase, and it stops when the player dies. I added bgm to the fight phase which stops when the player wins or dies.