added sfx for pickup items, closing walls
I added the sfx for picking up any items that are not boost items (i.e. weapons, key). There will be a sound upon pickup which lets the player know they picked it up when the weapon is instantiated in their inventory. I also added the sfx for the walls closing behind the player as they move through phases. I've adjusted volume of a few different sounds within Unity.